interactions/requirements.txt
# Required for Discord interaction request signature verification (Ed25519)
PyNaCl>=1.5.0
requests>=2.28.0
gspread>=5.12.0
oauth2client>=4.1.3
beautifulsoup4>=4.12.0
wiktionaryparser>=0.0.7
ety>=1.4.0
